Hoist SellAmountInput logic into SellPanel
